+++ CSFF_Mod_Framework/Plugin.cs
@@ -5,7 +5,7 @@ public class Plugin : BaseUnityPlugin
 {
     public const string PluginGuid = "crispywhips.CSFFModFramework";
     public const string PluginName = "CSFF Mod Framework";
-    public const string PluginVersion = "2.26.2";
+    public const string PluginVersion = "2.26.4";
 
     public static Plugin Instance { get; private set; }
     internal new static ManualLogSource Logger { get; private set; }